Struct transmog_bincode::Bincode

pub struct Bincode { /* fields omitted */ }

Bincode implementor of Format with default options.

Implementations

impl Bincode

pub fn legacy_default() -> Self

Returns a Bincode instance initialized using the equivalent settings that bincode::serialize, bincode::deserialize, bincode::serialize_into, and bincode::deserialize_from use. See bincode::config for more information.

pub fn no_limit(self) -> Self

Configures no byte limit. See Infinite for more information.

pub fn limit(self, byte_limit: u64) -> Self

Configures bincode to restrict encoding and decoding to byte_limit. See Bounded for more information.

pub fn big_endian(self) -> Self

Configures big-endian encoding. See BigEndian for more information.

pub fn little_endian(self) -> Self

Configures little-endian encoding. See LittleEndian for more information.

pub fn native_endian(self) -> Self

Configures native-endian encoding. See NativeEndian for more information.

pub fn variable_integer_encoding(self) -> Self

Configures variable length integer encoding. See VarintEncoding for more information.

pub fn fixed_integer_encoding(self) -> Self

Configures fixed length integer encoding. See FixintEncoding for more information.

pub fn allow_trailing_bytes(self) -> Self

Configures Bincode to allow trailing bytes when deserializing. See AllowTrailing for more information.

pub fn reject_trailing_bytes(self) -> Self

Configures Bincode to reject trailing bytes when deserializing. See RejectTrailing for more information.